+++ go/src/koding/klientctl/ctlcli/ctlcli.go
@@ -1,10 +1,7 @@
 // ctlcli holds the interfaces and helpers for the current CLI library
 // (codeganster/cli). Abstracting away any cli library implementation from the
 // commands themselves, keeping our commands testable and generic.
-//
-// TODO: Move this to it's own package, once the rest of the klientctl lib is
-// package-able.
-package main
+package ctlcli
 
 import (
 	"io"
